Bunch of cleanup and (hopefully) simplification
