+++ src/ParquetFileViewer.sln
@@ -1,11 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29201.188
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33213.308
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParquetFileViewer", "ParquetFileViewer\ParquetFileViewer.csproj", "{6019FC1B-3610-4682-BF96-8345C95CB7EC}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Utilities", "Utilities\Utilities.csproj", "{F423D115-06A0-47AF-A86E-2775E2F894F8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ParquetFileViewer", "ParquetFileViewer\ParquetFileViewer.csproj", "{6019FC1B-3610-4682-BF96-8345C95CB7EC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +15,6 @@ Global
 		{6019FC1B-3610-4682-BF96-8345C95CB7EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6019FC1B-3610-4682-BF96-8345C95CB7EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6019FC1B-3610-4682-BF96-8345C95CB7EC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F423D115-06A0-47AF-A86E-2775E2F894F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F423D115-06A0-47AF-A86E-2775E2F894F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F423D115-06A0-47AF-A86E-2775E2F894F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F423D115-06A0-47AF-A86E-2775E2F894F8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
