+++ src/ParquetFileViewer/MetadataViewer.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Windows.Forms;
-using Utilities;
+using ParquetFileViewer.Helpers;
 
 namespace ParquetFileViewer
 {
